==> quapos-lm/02-feature-extraction/02-cpfl1/01-a-feature-extraction-simpleitk.ipynb <==
## Feature extraction Cpfl1 - napari_simpleitk_image_processing

The dataset contains Cpfl1 animals and corresponding WT animals of the postnatal ages: 8, 14, 20, 30, 70, and 245.

The workflow for  extracting features with `napari_simpleitk_image_processing` and its `label_statistics` function was already demonstrated [here]. 

This notebook aims to extract features from all images in the corresponding folder using a for loop.

import apoc
import numpy as np
import os
import pandas as pd
import pyclesperanto_prototype as cle

from napari_simpleitk_image_processing import label_statistics

import sys
sys.path.append("../..")
from quapos_lm import rescale_image, rescale_segmentation, predict_image

# Load the classifier
quapos_lm = apoc.ObjectSegmenter(opencl_filename = "../../01-training-and-validation/02-quapos-lm.cl")
quapos_lm.feature_importances()

# Define file path for images
images = "../../data/02-data-for-pixel-classifier/cpfl-wt-comparison/b-s-ops-images-cropped/"

# Define a file list from the folder
file_list = os.listdir(images)
print(file_list)

### Extract features

Now a for loop will be computed to extract all the features. Additionally, information from the filename will be used to add the columns `age`, `genotype`, and `image_id`.

# Define empty array to store data
features = []

# Loop over the image folder
for i, file_name in enumerate(file_list):
    
    # Load the image 
    image = cle.imread(images + file_name)
    
    # Predict the image
    prediction = predict_image(image=image, classifier=quapos_lm)
    
    # Rescale original image
    image_rescaled = rescale_image(image=image, voxel_x=0.323, voxel_y=0.323, voxel_z=0.490)
    
    # Rescale the prediction
    prediction_rescaled = rescale_segmentation(segmentation=prediction, voxel_x=0.323, voxel_y=0.323, voxel_z=0.490)
    
    # Extract features with napari_simpleitk_image_processing's label_statistics function
    # Here all possible features are extracted
    features_i = label_statistics(
        intensity_image = image_rescaled,
        label_image = prediction_rescaled,
        size = True,
        intensity = True,
        perimeter = True,
        shape = True,
        position = True,
        moments = True)
    
    # Add information from the filename into corresponding columns
    # Split the filename
    file_name_split = file_name.split("-")
    
    # Add age in a respective column
    age = int(file_name_split[2].replace("p", ""))
    features_i["age"] = pd.Series([age for x in range(len(features_i))])
    
    # Add the image_id in a respective column
    features_i["image_id"] = pd.Series([i for x in range(len(features_i))])
    
    # Add the genotype in a respective column
    features_i["genotype"] = pd.Series([file_name_split[1] for x in range(len(features_i))])
    
    # Store measurements of current image in pandas dataframe
    features.append(features_i)
    
# Concatenate dataframe with all measurements
features = pd.concat(features)

# Show the dataframe
features

features.to_csv("../../measurements/01-cpfl-simpleitk.csv", index=False)
