==> quapos-lm/02-feature-extraction/02-cpfl1/02-combine-dataframes.ipynb <==
## Combine dataframes

Features were extracted from all images with [napari_simpleitk_image_processing](01-a-feature-extraction-simpleitk.ipynb), and [porespy](01-b-feature-extraction-porespy.ipynb). Now the two dataframes are combined on the index.

import pandas as pd
import numpy as np

# Define path were data is stored
path = "../../measurements/cpfl/"

# Load simpleitk measurements
features_simpleitk = pd.read_csv(path + "01-a-cpfl-simpleitk.csv")
features_simpleitk

# Load porespy measurements
features_porespy = pd.read_csv(path + "01-b-cpfl-porespy.csv")
features_porespy

### Combine dataframes

Dataframes are concatenated on their index.

# Combine dataframes
measurements = pd.concat([features_simpleitk, features_porespy], axis = 1)
measurements

measurements.to_csv(path + "02-features-combined.csv", index=False)
