==> quapos-lm/02-feature-extraction/02-cpfl1/03-remove-infinity-values.ipynb <==
## Remove infinity values

Similarly to the WT dataset, infinity values are present in the Cpfl1 dataset. Here these infinity values are removed. Simulltaneously, only relevant features are filtered.

import pandas as pd
import numpy as np

# Define path with measurements
path = "../../measurements/cpfl/"

# Load combined dataframe
measurements = pd.read_csv(path + "02-features-combined.csv")
measurements

# Filter for relevant features
measurements_filtered = measurements[['label', 'maximum', 'mean', 'median', 'minimum', 'sigma', 'sum',
       'variance', 'bbox_3', 'bbox_4', 'bbox_5', 'elongation',
       'feret_diameter', 'flatness', 'equivalent_ellipsoid_diameter_0', 'equivalent_ellipsoid_diameter_1',
       'equivalent_ellipsoid_diameter_2', 'equivalent_spherical_perimeter',
       'equivalent_spherical_radius', 'volume', 'surface_area', 'bbox_volume', 'convex_volume', 'sphericity',
       'solidity', 'perimeter_2d', 'major_axis_length_2d',
       'minor_axis_length_2d', 'age', 'image_id', 'genotype']]
measurements_filtered

### Remove infinity values

From the filtered dataframe, infinity values are removed.

# Check if inf values are present
# Command dosen't work when row contains string, genotype here is dropped
np.isinf(measurements_filtered.drop(labels="genotype", axis=1).values).any()

# Check how many rows contain infs
(measurements_filtered.isin([np.inf, -np.inf])).any(axis=1).sum()

# Remove rows with inf values
measurements_without_inf = measurements_filtered[~measurements_filtered.isin([np.inf, -np.inf]).any(axis=1)]
measurements_without_inf

measurements_without_inf.to_csv(path + "03-features-cpfl-inf-reduced.csv", index=False)
