==> quapos-lm/02-feature-extraction/02-cpfl1/04-size-rescaled.ipynb <==
## Rescale size features

The dataframe contains different size parameters, which require to be rescaled into their actual micrometer values from pixels/voxels.

import numpy as np
import pandas as pd

# Define path with measurements
path = "../../measurements/cpfl/"

# Load the measurements
measurements = pd.read_csv(path + "03-features-cpfl-inf-reduced.csv")
measurements.keys()

### Rescaling of 1D features

Features which only have 1 dimension are recalculated by multypling them with 0.323.

# Define 1D features
features_1d = ['bbox_3', 'bbox_4', 'bbox_5', 'equivalent_ellipsoid_diameter_0',
       'equivalent_ellipsoid_diameter_1', 'equivalent_ellipsoid_diameter_2',
       'equivalent_spherical_perimeter', 'equivalent_spherical_radius', 'feret_diameter', 'perimeter_2d', 'major_axis_length_2d',
       'minor_axis_length_2d']

# Apply recalculation for all features
for feature in features_1d:
    
    # Multiply them by 0.323
    measurements[feature + "_rescaled"] = measurements[feature] * 0.323
    
measurements

### Rescaled 2D feature

There is 1 feature which is 2D (surface area). It is recalculated into actual pixel values by multiplication with $0.323^2$

# Rescale the surface area, multiply by 0.323^2
measurements["surface_area_rescaled"] = measurements["surface_area"] * np.power(0.323, 2)
measurements

### Rescale 3D features

Certain features are 3 dimensional. They are rescaled by multiplying them with $0.323^3$.

# Define 3D features
features_3d = ['bbox_volume', 'convex_volume', 'volume']

# Apply for loop to rescale those features
for feature in features_3d:
    
    # Rescale and add "_rescaled" to the name of the columns
    measurements[feature + "_rescaled"] = measurements[feature] * np.power(0.323, 3)
    
measurements

### Filter duplicated features

All rescaled features are now duplicated. Non-rescaled size features will be removed.

# Keep only rescaled features
measurements_rescaled = measurements[['label', 'age', 'image_id', 'genotype', 'maximum', 'mean', 'median', 'minimum', 'sigma', 'sum',
       'variance', 'flatness', 'elongation','sphericity', 'solidity', 'bbox_3_rescaled', 'bbox_4_rescaled', 'bbox_5_rescaled',
       'equivalent_ellipsoid_diameter_0_rescaled',
       'equivalent_ellipsoid_diameter_1_rescaled',
       'equivalent_ellipsoid_diameter_2_rescaled',
       'equivalent_spherical_perimeter_rescaled',
       'equivalent_spherical_radius_rescaled', 'feret_diameter_rescaled',
       'perimeter_2d_rescaled', 'major_axis_length_2d_rescaled',
       'minor_axis_length_2d_rescaled', 'surface_area_rescaled',
       'bbox_volume_rescaled', 'convex_volume_rescaled', 'volume_rescaled']]
measurements_rescaled

measurements_rescaled.to_csv(path + "04-measurements-rescaled.csv", index=False)
