==> quapos-lm/02-feature-extraction/02-cpfl1/05-filter-5-and-95-percentiles.ipynb <==
## Filter dataset based on 5- and 95- percentiles

The classifier still predicts labels which correspond to several outer segments or which are too small to be considered a outer segment at all.

Here a filter of the 5- and 95- percentiles are applied based on the rescaled volume of the outer segments. 

Here, the 2 genotypes are handled independently from each other. Meaning, the dataframe will be split into a WT and a Cpfl1 dataset. Then, the 5- and 95- percentiles are calculated for both dataframes. Afterwards, the two dataframes will be filtered.

At the end both dataframes will be combined again.

import numpy as np
import pandas as pd

# Define the path were measurements are stored
path = "../../measurements/cpfl/"

# Load measurements
measurements = pd.read_csv(path + "04-measurements-rescaled.csv")
measurements

### Filter WT dataset

First, the WT dataset will be filtered.

# Retrieve WT measurements
measurements_wt = measurements[measurements["genotype"] == "wt"]
measurements_wt

Compute 5- and 95- percentiles of the WT dataset.

# Compute 5- percentile of WT
five_percentile_wt = measurements_wt["volume_rescaled"].quantile(0.05)
five_percentile_wt

# Compute 95- percentiles of the WT dataset
ninety_five_percentile_wt = measurements_wt["volume_rescaled"].quantile(0.95)
ninety_five_percentile_wt

Now, the dataset will be filtered based on those percentiles and the rescaled volume.

# Filter wt measurements
measurements_wt_filtered = measurements_wt[(measurements_wt["volume_rescaled"] > five_percentile_wt) &
                                           (measurements_wt["volume_rescaled"] < ninety_five_percentile_wt)]
measurements_wt_filtered

### Filter Cpfl1 dataset

Second, the Cpfl1 dataset will be filtered.

# Retrieve Cpfl1 dataset
measurements_cpfl = measurements[measurements["genotype"] == "cpfl"]
measurements_cpfl

Compute 5- and 95- percentile of the Cpfl1 dataset.

# Calculate five percentile of Cpfl1 dataset
five_percentile_cpfl = measurements_cpfl["volume_rescaled"].quantile(0.05)
five_percentile_cpfl

# Calculate 95- percentile of Cpfl1 dataset
ninety_five_percentile_cpfl = measurements_cpfl["volume_rescaled"].quantile(0.95)
ninety_five_percentile_cpfl

Apply filters to Cpfl1 dataset, volume rescaled.

# Filter volume of Cpfl1 dataset
measurements_cpfl_filtered = measurements_cpfl[(measurements_cpfl["volume_rescaled"] > five_percentile_cpfl) &
                                               (measurements_cpfl["volume_rescaled"] < ninety_five_percentile_cpfl)]
measurements_cpfl_filtered

### Combine filtered dataframes

Both dataframes were filtered and can be combined now.

# Combine dataframes
measurements_filtered = pd.concat([measurements_cpfl_filtered, measurements_wt_filtered], ignore_index=True)
measurements_filtered

measurements_filtered.to_csv(path + "05-measurements-filtered-5-and-95-percentiles.csv", index=False)
