==> quapos-lm/02-feature-extraction/02-cpfl1/06-average-values.ipynb <==
## Compute average values for statistics

For the subsequent statistics, the average values of the features should be plotted. Here a dataframe will be computed which contains average values from the filtered dataframe.

For the Cpfl1 dataset only 1 technical replicate was computed. Hence here the computation only needs to be done by the `image_id`.

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/cpfl/"

# Load measurements
measurements = pd.read_csv(path + "05-measurements-filtered-5-and-95-percentiles.csv")
measurements

### Average per image id

Here the values are averaged by `image_id`.

# Compute average values per image id
measurements_means = measurements.groupby("image_id", as_index=False).mean()
measurements_means

### Filter column: `label`

For the label instead of the mean number of label per image, the max number of labels per image should be determined. This will be done in the next notebook.

# Drop the label column
measurements_means = measurements_means.drop(labels=["label"], axis=1)
measurements_means

measurements_means.to_csv(path + "06-average-values.csv", index=False)
