==> quapos-lm/02-feature-extraction/02-cpfl1/07-add-max-number-of-labels.ipynb <==
## Calculate max number of labels

In the [previous notebook](06-average-values.ipynb) average values of the features (per image and per biological replicate) were calculated for creating plots and statistics subsequently. Due to that the number of labels were misleading (should be maximum per image instead of average per image). 

This notebook computes the maximum number of labels from the filtered dataset and adds them to the average dataframe.

Additionally, while computing the average values the `genotype` column was removed as well, it will be added here again.

import numpy as np
import pandas as pd

# Define the path were measurements are stored
path = "../../measurements/cpfl/"

# Load filtered measurements
measurements = pd.read_csv(path + "05-measurements-filtered-5-and-95-percentiles.csv")
measurements

### Determine max number of labels per image

Now the max number of labels per image are determined with the `df.groupby()` function. The argument `.size()` will be used instead of `.max()` since certain rows were filtered from the dataframe. This action requires to rename the resulting column from "size" into "label" again.

# Determine max number of labels
number_of_labels = measurements.groupby("image_id", as_index=False)["label"].size()

# Rename resulting column
number_of_labels.rename(columns={"size": "label"}, inplace=True)
number_of_labels

### Load averaged dataframe

Now, the averaged dataframe is loaded and the number of labels added to it.

# Load mean dataframe
measurements_means = pd.read_csv(path + "06-average-values.csv")

# Add number of labels
measurements_means = pd.merge(measurements_means, number_of_labels, how="inner")
measurements_means

### Retrieve genotype and add them again to dataframe

Additionally, the genotype was lost while calculating averages, will be added here again. To calculate the genotype grouped by the `image_id` the `.min()` argument can be passed on to the `.groupby()` function.

# Retrieve genotype
genotype = measurements.groupby(["image_id"], as_index=False)["genotype"].min()
genotype

Add genotype to averaged dataframe.

# Add genotype to dataframe
measurements_means = pd.merge(measurements_means, genotype, how="inner")
measurements_means

measurements_means.to_csv(path + "07-average-values-max-number-of-labels-added.csv", index=False)
