==> quapos-lm/02-feature-extraction/02-cpfl1/08-add-summed-volume.ipynb <==
## Summed OS volume

One way to circumvent that several outer segments are combined in one label (without filtering) is calculating the summed OS volume per image and afterwards the average per biological replicate.

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/cpfl/"

# Load unfiltered measurements
measurements = pd.read_csv(path + "04-measurements-rescaled.csv")
measurements

### Calculate the summed volume per image_id

Here, the summed volume per image id is calculated.

# Calculate summed OS volume
summed_volume = measurements.groupby("image_id", as_index=False)["volume_rescaled"].sum()

# Rename column for better identification
summed_volume.rename(columns={"volume_rescaled": "summed_volume"}, inplace=True)
summed_volume

### Open average dataframe and add summed volume

# Load mean dataframe
measurements_means = pd.read_csv(path + "07-average-values-max-number-of-labels-added.csv")

# Add summed volume
measurements_processed = pd.merge(measurements_means, summed_volume, how="inner")
measurements_processed["age"] = measurements_processed["age"].astypeype(int)
measurements_processed

measurements_processed.to_csv(path + "08-features-processed.csv", index=False)
