==> quapos-lm/02-feature-extraction/02-cpfl1/09-rename-columns-for-correlation-matrix.ipynb <==
## Rename columns for correlation matrix

For better visualisation of the correlation matrix the column names are renamed. This include:

- turning bounding box 3, 4, and 5 to their actual width, height, and depth parameters
- turning equivalent ellipsoid diameter 0, 1, and 2 into their actual height, width, and depth parameters
- removing the "rescaled" from the size parameteres
- properly label intensity parameters

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/cpfl/"

# Load dataset
measurements = pd.read_csv(path + "08-features-processed.csv")
measurements.keys()

### Rename the features

To rename the features a dictionary will be defined.

# Define dictionary to rename columns
column_names = {"label": "label",
                "age": "age",
                "maximum": "max_intensity",
                "mean": "mean_intensity",
                "median": "median_intensity",
                "minimum": "minimum_intensity",
                "sigma": "sigma_intensity",
                "sum": "sum_intensity",
                "variance": "variance_intensity",
                "bbox_3_rescaled": "bbox_width",
                "bbox_4_rescaled": "bbox_height",
                "bbox_5_rescaled": "bbox_depth",
                "equivalent_ellipsoid_diameter_0_rescaled": "equivalent_ellipsoid_diameter_depth",
                "equivalent_ellipsoid_diameter_1_rescaled": "equivalent_ellipsoid_diameter_width",
                "equivalent_ellipsoid_diameter_2_rescaled": "equivalent_ellipsoid_diameter_height",
                "equivalent_spherical_perimeter_rescaled": "equivalent_spherical_perimeter",
                "equivalent_spherical_radius_rescaled": "equivalent_spherical_radius",
                "feret_diameter_rescaled": "feret_diamter",
                "perimeter_2d_rescaled": "perimeter_2d",
                "major_axis_length_2d_rescaled": "major_axis_length_2d",
                "minor_axis_length_2d_rescaled": "minor_axis_length_2d",
                "surface_area_rescaled": "surface_area",
                "bbox_volume_rescaled": "bbox_volume",
                "volume_rescaled": "volume",
                "summed_volume_mean": "summed_volume"}

# Rename the columns
measurements.rename(columns=column_names, inplace=True)
measurements.keys()

measurements.to_csv(path + "09-columns-renamed.csv", index=False)
